+++ src/events/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Discord Debate Timer
+ *
+ * @license BSD-3-Clause
+ * @author Luke Zhang luke-zhang-04.github.io/
+ * @copyright 2020 - 2021 Luke Zhang
+ */
+
+export {onVoiceStateUpdate} from "./onVoiceStateUpdate"